+++ velocity/src/main/java/io/tebex/plugin/command/sub/DebugCommand.java
@@ -5,7 +5,10 @@ import dev.dejvokep.boostedyaml.YamlDocument;
 import io.tebex.plugin.TebexPlugin;
 import io.tebex.plugin.command.SubCommand;
 import io.tebex.sdk.platform.config.ProxyPlatformConfig;
+import io.tebex.sdk.util.StringUtil;
 import net.kyori.adventure.text.Component;
+
+import java.io.IOException;
 
 public class DebugCommand extends SubCommand {
     public DebugCommand(TebexPlugin platform) {
@@ -19,15 +22,22 @@ public class DebugCommand extends SubCommand {
         ProxyPlatformConfig config = platform.getPlatformConfig();
         YamlDocument configFile = config.getYamlDocument();
 
-        boolean enableDebug = Boolean.parseBoolean(args[0]);
-        if (enableDebug) {
+        if (StringUtil.isTruthy(args[0])) {
             sender.sendMessage(Component.text("§b[Tebex] §7Debug mode enabled."));
             config.setVerbose(true);
             configFile.set("verbose", true);
-        } else {
+        } else if (StringUtil.isFalsy(args[0])){
             sender.sendMessage(Component.text("§b[Tebex] §7Debug mode disabled."));
             config.setVerbose(false);
             configFile.set("verbose", false);
+        } else {
+            sender.sendMessage(Component.text("§b[Tebex] §7Invalid command usage. Use /tebex " + this.getName() + " " + getUsage()));
+        }
+
+        try {
+            configFile.save();
+        } catch (IOException e) {
+            sender.sendMessage(Component.text("§b[Tebex] §7Failed to save configuration file."));
         }
     }
 
@@ -38,6 +48,6 @@ public class DebugCommand extends SubCommand {
 
     @Override
     public String getUsage() {
-        return "<trueOrFalse>";
+        return "<true/false/on/off>";
     }
 }